+++ src/QNEthernetUDP.h
@@ -118,6 +118,13 @@ class EthernetUDP : public UDP {
   IPAddress remoteIP() final;
   uint16_t remotePort() final;
 
+  // Returns the approximate packet arrival time, measured with millis(). This
+  // is only valid if a packet has been received with parsePacket().
+  //
+  // This is useful in the case where packets have been queued and the caller
+  // needs the approximate arrival time.
+  uint32_t timestamp() const;
+
   // Returns whether the socket is listening.
   explicit operator bool() const;
 
@@ -126,6 +133,7 @@ class EthernetUDP : public UDP {
     std::vector<uint8_t> data;
     ip_addr_t addr = *IP_ANY_TYPE;
     volatile uint16_t port = 0;
+    volatile uint32_t timestamp = 0;  // Approximate arrival time
   };
 
   static void recvFunc(void *arg, struct udp_pcb *pcb, struct pbuf *p,